refactor(src): seperate out truely 'core' code from 'jscad' - makes it more clear what is a piece of code used for cli, module & browser - other input/output related code moved out to io folder - some minor cleanup & docstring work - modified all import paths & ran tests to verify coherence
