+++ src/core/jscad-function.js
@@ -1,8 +1,3 @@
-// == OpenJSCAD.org, Copyright (c) 2013-2016, Licensed under MIT License
-//
-// History:
-//   2016/02/02: 0.4.0: GUI refactored, functionality split up into more files, mostly done by Z3 Dev
-
 /**
  * Create an function for processing the JSCAD script into CSG/CAG objects
  * @param {String} script the script
