+++ src/core/rebuildSolid.js
