+++ src/core/replaceIncludes.js
