+++ src/core/resolveIncludes.js
