+++ src/utils/generateOutputData.js
@@ -1,7 +1,7 @@
 import oscad from '@jscad/scad-api'
 import { convertToBlob } from '../io/convertToBlob'
 import { formats } from '../io/formats'
-import { rebuildSolid } from '../jscad/rebuildSolid'
+import { rebuildSolid } from '../core/rebuildSolid'
 import { resolveIncludesFs } from './resolveIncludesFs'
 import getParameterDefinitionsCLI from './getParameterDefinitionsCLI'
